Adds a bunch of provisional setup for local sql and deployed db, but not working yet
