+++ Pipfile
@@ -12,8 +12,7 @@ url = "https://pypi.python.org/simple"
 [packages]
 
 gunicorn = "*"
-flask = "*"
-flask-sqlalchemy = "*"
+Flask = "*"
+Flask-SQLAlchemy = "*"
 "psycopg2" = "*"
 flask-migrate = "*"
-flask-script = "*"
