+++ Procfile
@@ -1,4 +1,4 @@
 web: gunicorn app:app
-init: python app/manage.py db init
-migrate: python app/manage.py db migrate
-upgrade: python app/manage.py db upgrade
+init: flask db init
+migrate: flask db migrate
+upgrade: flask db upgrade
