--- app/app.py
@@ -1,30 +1,0 @@
-import os
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-
-app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = os.environ["DATABASE_URL"]
-db = SQLAlchemy(app)
-
-from app.models import User
-
-@app.route('/')
-def hello_world():
-    return 'Hello, World!'
-
-@app.route('/add/')
-def webhook():
-    name = "test"
-    email = "test@olin.edu"
-    u = User(id = id, name = name, email = email)
-    print("user created", u)
-    db.session.add(u)
-    db.session.commit()
-    return "user created"
-
-@app.route('/delete/')
-def delete():
-    u = User.query.get(i)
-    db.session.delete(u)
-    db.session.commit()
-    return "user deleted"
