--- app/manage.py
@@ -1,13 +1,0 @@
-from flask_script import Manager
-from flask_migrate import Migrate, MigrateCommand
-
-from app.app import app, db
-
-migrate = Migrate(app, db)
-
-manager = Manager(app)
-manager.add_command('db', MigrateCommand)
-
-
-if __name__ == '__main__':
-    manager.run()
