+++ notes.md
@@ -1,0 +1,1 @@
+In order for flask CLI to work, `FLASK_APP` env variable has to be set to app.
